Add files via upload
The google location history data dump gives the location of the user, i.e., the latitude and longitude details, at each time stamp. This data has to be smoothened to handle the measurement errors and it the necessary fields had to be created which would shape the further analysis. 
The entire data cleanup and data wrangling was done using RStudio. 
The first step is to calculate the distance travelled and the associated velocities, for each pair of time stamps. 
Once these are computed, we can proceed to the pattern recognition aspect of the analysis. Which is the most important part of the analysis, to arrive at the timestamps when the user was not stationary i.e., when the user was undertaking a journey. 
Once the journeys are identified, the similar ones can be clustered, so that time saved if those journeys were made using a car/own transport, can be calculated.